requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-redis==6.4.0
redis==6.4.0
psycopg2-binary==2.9.10
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-cors-headers==4.6.0
django-filter==24.3
python-decouple==3.8
